Allow non-0-based indices, and remove constraints (#10)
* Fisher-Yates shuffle: allow non-0-based indices and require random-access

Fixed two issues:

1. shuffle() will no longer fail when used on collections whose indices are not zero-based (eg. ArraySlice).

2. shuffle() now requires the collection to be random-access, which was previously an unstated assumption.

Requiring random-access also improves the running time from quadratic to linear, because “count” inside the loop is now constant-time rather than potentially O(n).

* Fixed malformed conditional compilation blocks

Fixed the conditional compilation blocks so they actually work now.

* Fixed typo

A couple of characters got lost in the last commit, added them back.

* Eliminate all mention of IndexDistance

The implementation no longer constrains IndexDistance, and the first “#if” block no longer requires swift(>=2.2).

Also added “swapAt” for older Swift versions, to simplify the inner loop of ”shuffle”.

* Removed random-access constraint for shuffle()

The complexity of “shuffle()” is O(n) for random-access collections, and O(n^2) otherwise.

Note that “shuffled()” is always O(n).

* Made random() generic, and fixed modulo bias

To accommodate the removal of the “IndexDistance == Int” constraint in shuffle(), the random() function is now generic over SignedInteger.

Also, the Glibc version of random() should no longer exhibited modulo bias.

* Fixed generic constraints

We need “random” to be generic over SignedInteger (at a minimum) because that is the constraint on Collection.IndexDistance which is our use case in “shuffle”. However it would be strange if “random” did not to work for unsigned integers, so we allow those too.

It is unfortunate that Swift 3.1 had a BinaryInteger protocol which SignedInteger did not conform to, since otherwise we could spell the typealias “BinaryInteger”.

* Fixed Swift version tests

Had some silly mistakes in the Swift version tests. Fixed now.

* Made “random” a function instead of a closure

The “random” function is now a function rather than a closure constant.

* Guard against infinite loop

Added a guard statement to prevent an infinite loop in the Glibc version of “random” when the input is negative.

* Added precondition(n > 0) to Glibc random

Replaced “guard” with “precondition”.

* Removed Swifter version checks

Since we are only trying to support Swift 4+, all the conditional compilation for older Swift versions is now gone.

* Removed Swift version checks

Since we are only trying to support Swift 4+, all the conditional compilation for older Swift versions is now gone.

Also improved shuffle() performance on non-random-access collections. Complexity remains the same: O(n) with random access and O(n^2) without.

* Swapped 2 lines

Minor change to make updating the loop condition be the last thing in the loop.

* Added missing "return"

A "return" keyword was missing.